refactor: standardize tender step management using unified hooks and refactored section components across show pages
